--- src/containers/p2p/flowPayment/style.css
@@ -1,97 +1,0 @@
-:local .modalBox {
-    width: 100%;
-    text-align: center;
-    padding-top: 30px;
-    padding-bottom: 30px;
-    color: #fff;
-  }
-  :local .modalIconCoin {
-    max-width: 50px;
-    margin-bottom: 20px;
-  }
-  :local .iconInfor {
-    max-width: 125px;
-    max-height: 125px;
-    margin-bottom: 20px;
-  }
-  :local .totalConfirm {
-    font-size: 18px;
-    font-weight: 700;
-  }
-  :local .addressConfirm {
-    composes: clGreenDefault from "../../../components/bases/colorBase.css";
-    font-weight: 700;
-  }
-  :local .boxReceive .address{
-    color: #68f285;
-    text-align: center;
-  }
-  :local .confirmFee {
-    margin-top: 40px;
-    margin-bottom: 40px;
-    font-size: 10px;
-  }
-  :local .txtamount {
-    composes: inputTextDefault from "../../../components/bases/inputBase.css";
-    background: transparent;
-    font-size: 42px;
-    text-align: center;
-  }
-  
-  :local .txtamount::placeholder {
-    font-size: 42px;
-    text-align: center;
-  }
-  :local .boxFee {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    width: 100%;
-    margin-bottom: 10px;
-  }
-  :local .labelFee {
-    display: block;
-    padding: 5px;
-    padding-left: 10px;
-    padding-right: 10px;
-    margin-left: 10px;
-    margin-right: 10px;
-    font-size: 12px;
-    cursor: pointer;
-    opacity: 0.8;
-    transition: 0.3s;
-  }
-  
-  :local .labelFee:hover {
-    opacity: 1;
-    border-bottom-width: 3px;
-    padding-top: 3px;
-  }
-  
-  :local .greenLabelFee {
-    composes: labelFee;
-    border-bottom: solid 1px greenyellow;
-  }
-  :local .yellowLabelFee {
-    composes: labelFee;
-    border-bottom: solid 1px orange;
-  }
-  :local .redLabelFee {
-    composes: labelFee;
-    border-bottom: solid 1px red;
-  }
-  :local .paddingTop8 {
-    padding-top: 8%;
-  }
-  :local .btGreen{
-    composes: buttonBorderGreen from "../../../components/bases/buttonBase.css";
-  }
-  @media only screen and (max-width: 400px) {
-    :local .modalBox {
-      font-size: 12px;
-    }
-    :local .baseStep {
-      padding: 30px 0;
-    }
-  }
-
